pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "community-scraper"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.27",
    "trafilatura>=1.12",
    "html2text>=2024.2",
    "pydantic>=2.7",
    "pyyaml>=6.0",
    "APScheduler>=3.10",
    "structlog>=24.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "respx>=0.21",
    "ruff>=0.4",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["scraper*"]

[tool.ruff]
line-length = 100
target-version = "py312"
